move file into views folder
